Move router test utils into their own module
Useful for testing onion message pathfinding in upcoming PRs and any future
routing modules we add
