+++ lightning/src/routing/mod.rs
@@ -12,3 +12,5 @@
 pub mod gossip;
 pub mod router;
 pub mod scoring;
+#[cfg(test)]
+mod test_utils;
